iterate on IntUiBridgeProgress
