+++ ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridge.kt
@@ -258,27 +258,6 @@ private fun readGroupHeaderStyle() =
             ),
     )
 
-private fun readHorizontalProgressBarStyle() =
-    HorizontalProgressBarStyle(
-        colors =
-            HorizontalProgressBarColors(
-                track = retrieveColorOrUnspecified("ProgressBar.trackColor"),
-                progress = retrieveColorOrUnspecified("ProgressBar.progressColor"),
-                indeterminateBase = retrieveColorOrUnspecified("ProgressBar.indeterminateStartColor"),
-                indeterminateHighlight = retrieveColorOrUnspecified("ProgressBar.indeterminateEndColor"),
-            ),
-        metrics =
-            HorizontalProgressBarMetrics(
-                cornerSize = CornerSize(100),
-                minHeight = 4.dp, // See DarculaProgressBarUI.DEFAULT_WIDTH
-                // See DarculaProgressBarUI.CYCLE_TIME_DEFAULT,
-                // DarculaProgressBarUI.REPAINT_INTERVAL_DEFAULT,
-                // and the "step" constant in DarculaProgressBarUI#paintIndeterminate
-                indeterminateHighlightWidth = (800 / 50 * 6).dp,
-            ),
-        indeterminateCycleDuration = 800.milliseconds, // See DarculaProgressBarUI.CYCLE_TIME_DEFAULT
-    )
-
 private fun readLinkStyle(): LinkStyle {
     val normalContent =
         retrieveColorOrUnspecified("Link.activeForeground").takeOrElse {
